Update operation docs
