+++ cmd/broker/main.go
@@ -43,15 +43,9 @@ import (
 	"github.com/kyma-project/kyma-environment-broker/internal/notification"
 	"github.com/kyma-project/kyma-environment-broker/internal/orchestration"
 	orchestrate "github.com/kyma-project/kyma-environment-broker/internal/orchestration/handlers"
-	"github.com/kyma-project/kyma-environment-broker/internal/orchestration/manager"
 	"github.com/kyma-project/kyma-environment-broker/internal/process"
-	"github.com/kyma-project/kyma-environment-broker/internal/process/deprovisioning"
 	"github.com/kyma-project/kyma-environment-broker/internal/process/input"
 	"github.com/kyma-project/kyma-environment-broker/internal/process/provisioning"
-	"github.com/kyma-project/kyma-environment-broker/internal/process/steps"
-	"github.com/kyma-project/kyma-environment-broker/internal/process/update"
-	"github.com/kyma-project/kyma-environment-broker/internal/process/upgrade_cluster"
-	"github.com/kyma-project/kyma-environment-broker/internal/process/upgrade_kyma"
 	"github.com/kyma-project/kyma-environment-broker/internal/provider"
 	"github.com/kyma-project/kyma-environment-broker/internal/provisioner"
 	"github.com/kyma-project/kyma-environment-broker/internal/reconciler"
@@ -169,10 +163,6 @@ type Config struct {
 	Profiler ProfilerConfig
 
 	Events events.Config
-
-	Provisioning   process.StagedManagerConfiguration
-	Deprovisioning process.StagedManagerConfiguration
-	Update         process.StagedManagerConfiguration
 }
 
 type ProfilerConfig struct {
@@ -625,410 +615,6 @@ func panicOnError(err error) {
 	if err != nil {
 		panic(err)
 	}
-}
-
-func NewProvisioningProcessingQueue(ctx context.Context, provisionManager *process.StagedManager, workersAmount int, cfg *Config,
-	db storage.BrokerStorage, provisionerClient provisioner.Client, inputFactory input.CreatorForPlan, avsDel *avs.Delegator,
-	internalEvalAssistant *avs.InternalEvalAssistant, externalEvalCreator *provisioning.ExternalEvalCreator,
-	runtimeVerConfigurator *runtimeversion.RuntimeVersionConfigurator,
-	runtimeOverrides provisioning.RuntimeOverridesAppender, edpClient provisioning.EDPClient, accountProvider hyperscaler.AccountProvider,
-	reconcilerClient reconciler.Client, k8sClientProvider func(kcfg string) (client.Client, error), cli client.Client, logs logrus.FieldLogger) *process.Queue {
-
-	const postActionsStageName = "post_actions"
-	provisionManager.DefineStages([]string{startStageName, createRuntimeStageName,
-		checkKymaStageName, createKymaResourceStageName, postActionsStageName})
-	/*
-			The provisioning process contains the following stages:
-			1. "start" - changes the state from pending to in progress if no deprovisioning is ongoing.
-			2. "create_runtime" - collects all information needed to make an input for the Provisioner request as overrides and labels.
-			Those data is collected using an InputCreator which is not persisted. That's why all steps which prepares such data must be in the same stage as "create runtime step".
-		    All steps which requires InputCreator must be run in this stage.
-			3. "check_kyma" - checks if the Kyma is installed
-			4. "post_actions" - all steps which must be executed after the runtime is provisioned
-
-			Once the stage is done it will never be retried.
-	*/
-
-	provisioningSteps := []struct {
-		disabled  bool
-		stage     string
-		step      process.Step
-		condition process.StepCondition
-	}{
-		{
-			stage: startStageName,
-			step:  provisioning.NewStartStep(db.Operations(), db.Instances()),
-		},
-		{
-			stage: createRuntimeStageName,
-			step:  provisioning.NewInitialisationStep(db.Operations(), db.Instances(), inputFactory, runtimeVerConfigurator),
-		},
-		{
-			stage: createRuntimeStageName,
-			step:  steps.NewInitKymaTemplate(db.Operations()),
-		},
-		{
-			stage:     createRuntimeStageName,
-			step:      provisioning.NewResolveCredentialsStep(db.Operations(), accountProvider),
-			condition: provisioning.SkipForOwnClusterPlan,
-		},
-		{
-			stage:     createRuntimeStageName,
-			step:      provisioning.NewEDPRegistrationStep(db.Operations(), edpClient, cfg.EDP),
-			disabled:  cfg.EDP.Disabled,
-			condition: provisioning.SkipForOwnClusterPlan,
-		},
-		{
-			stage: createRuntimeStageName,
-			step:  provisioning.NewOverridesFromSecretsAndConfigStep(db.Operations(), runtimeOverrides, runtimeVerConfigurator),
-			// Preview plan does not call Reconciler so it does not need overrides
-			condition: skipForPreviewPlan,
-		},
-		{
-			condition: provisioning.WhenBTPOperatorCredentialsProvided,
-			stage:     createRuntimeStageName,
-			step:      provisioning.NewBTPOperatorOverridesStep(db.Operations()),
-		},
-		{
-			condition: provisioning.SkipForOwnClusterPlan,
-			stage:     createRuntimeStageName,
-			step:      provisioning.NewCreateRuntimeWithoutKymaStep(db.Operations(), db.RuntimeStates(), db.Instances(), provisionerClient),
-		},
-		{
-			condition: provisioning.DoForOwnClusterPlanOnly,
-			stage:     createRuntimeStageName,
-			step:      provisioning.NewCreateRuntimeForOwnClusterStep(db.Operations(), db.Instances()),
-		},
-		{
-			stage:     createRuntimeStageName,
-			step:      provisioning.NewCheckRuntimeStep(db.Operations(), provisionerClient, cfg.Provisioner.ProvisioningTimeout),
-			condition: provisioning.SkipForOwnClusterPlan,
-		},
-		{
-			stage: createRuntimeStageName,
-			step:  provisioning.NewGetKubeconfigStep(db.Operations(), provisionerClient, k8sClientProvider),
-		},
-		{
-			condition: provisioning.WhenBTPOperatorCredentialsProvided,
-			stage:     createRuntimeStageName,
-			step:      provisioning.NewInjectBTPOperatorCredentialsStep(db.Operations(), k8sClientProvider),
-		},
-		{
-			disabled: cfg.LifecycleManagerIntegrationDisabled,
-			stage:    createRuntimeStageName,
-			step:     steps.SyncKubeconfig(db.Operations(), cli),
-		},
-		{
-			disabled:  cfg.ReconcilerIntegrationDisabled,
-			stage:     createRuntimeStageName,
-			step:      provisioning.NewCreateClusterConfiguration(db.Operations(), db.RuntimeStates(), reconcilerClient),
-			condition: skipForPreviewPlan,
-		},
-		{
-			disabled:  cfg.ReconcilerIntegrationDisabled,
-			stage:     checkKymaStageName,
-			step:      provisioning.NewCheckClusterConfigurationStep(db.Operations(), reconcilerClient, cfg.Reconciler.ProvisioningTimeout),
-			condition: skipForPreviewPlan,
-		},
-		{
-			disabled: cfg.LifecycleManagerIntegrationDisabled,
-			stage:    createKymaResourceStageName,
-			step:     provisioning.NewApplyKymaStep(db.Operations(), cli),
-		},
-		// post actions
-		{
-			stage: postActionsStageName,
-			step:  provisioning.NewExternalEvalStep(externalEvalCreator),
-		},
-		{
-			stage:    postActionsStageName,
-			step:     provisioning.NewInternalEvaluationStep(avsDel, internalEvalAssistant),
-			disabled: cfg.Avs.Disabled,
-		},
-	}
-	for _, step := range provisioningSteps {
-		if !step.disabled {
-			err := provisionManager.AddStep(step.stage, step.step, step.condition)
-			if err != nil {
-				fatalOnError(err)
-			}
-		}
-	}
-
-	queue := process.NewQueue(provisionManager, logs)
-	queue.Run(ctx.Done(), workersAmount)
-
-	return queue
-}
-
-func NewUpdateProcessingQueue(ctx context.Context, manager *process.StagedManager, workersAmount int, db storage.BrokerStorage, inputFactory input.CreatorForPlan,
-	provisionerClient provisioner.Client, publisher event.Publisher, runtimeVerConfigurator *runtimeversion.RuntimeVersionConfigurator, runtimeStatesDb storage.RuntimeStates,
-	runtimeProvider input.ComponentListProvider, reconcilerClient reconciler.Client, cfg Config, k8sClientProvider func(kcfg string) (client.Client, error), cli client.Client, logs logrus.FieldLogger) *process.Queue {
-
-	requiresReconcilerUpdate := update.RequiresReconcilerUpdate
-	if cfg.ReconcilerIntegrationDisabled {
-		requiresReconcilerUpdate = func(op internal.Operation) bool { return false }
-	}
-	manager.DefineStages([]string{"cluster", "btp-operator", "btp-operator-check", "check"})
-	updateSteps := []struct {
-		stage     string
-		step      process.Step
-		condition process.StepCondition
-	}{
-		{
-			stage: "cluster",
-			step:  update.NewInitialisationStep(db.Instances(), db.Operations(), runtimeVerConfigurator, inputFactory),
-		},
-		{
-			stage:     "cluster",
-			step:      update.NewUpgradeShootStep(db.Operations(), db.RuntimeStates(), provisionerClient),
-			condition: update.SkipForOwnClusterPlan,
-		},
-		{
-			stage: "btp-operator",
-			step:  update.NewInitKymaVersionStep(db.Operations(), runtimeVerConfigurator, runtimeStatesDb),
-		},
-		{
-			stage:     "btp-operator",
-			step:      update.NewGetKubeconfigStep(db.Operations(), provisionerClient, k8sClientProvider),
-			condition: update.ForBTPOperatorCredentialsProvided,
-		},
-		{
-			stage:     "btp-operator",
-			step:      update.NewBTPOperatorOverridesStep(db.Operations(), runtimeProvider),
-			condition: update.RequiresBTPOperatorCredentials,
-		},
-		{
-			stage:     "btp-operator",
-			step:      update.NewApplyReconcilerConfigurationStep(db.Operations(), db.RuntimeStates(), reconcilerClient),
-			condition: requiresReconcilerUpdate,
-		},
-		{
-			stage:     "btp-operator-check",
-			step:      update.NewCheckReconcilerState(db.Operations(), reconcilerClient),
-			condition: update.CheckReconcilerStatus,
-		},
-		{
-			stage:     "check",
-			step:      update.NewCheckStep(db.Operations(), provisionerClient, 40*time.Minute),
-			condition: update.SkipForOwnClusterPlan,
-		},
-	}
-
-	for _, step := range updateSteps {
-		err := manager.AddStep(step.stage, step.step, step.condition)
-		if err != nil {
-			fatalOnError(err)
-		}
-	}
-	queue := process.NewQueue(manager, logs)
-	queue.Run(ctx.Done(), workersAmount)
-
-	return queue
-}
-
-func NewDeprovisioningProcessingQueue(ctx context.Context, workersAmount int, deprovisionManager *process.StagedManager,
-	cfg *Config, db storage.BrokerStorage, pub event.Publisher,
-	provisionerClient provisioner.Client, avsDel *avs.Delegator, internalEvalAssistant *avs.InternalEvalAssistant,
-	externalEvalAssistant *avs.ExternalEvalAssistant, bundleBuilder ias.BundleBuilder,
-	edpClient deprovisioning.EDPClient, accountProvider hyperscaler.AccountProvider, reconcilerClient reconciler.Client,
-	k8sClientProvider func(kcfg string) (client.Client, error), cli client.Client, configProvider input.ConfigurationProvider, logs logrus.FieldLogger) *process.Queue {
-
-	deprovisioningSteps := []struct {
-		disabled bool
-		step     process.Step
-	}{
-		{
-			step: deprovisioning.NewInitStep(db.Operations(), db.Instances(), 12*time.Hour),
-		},
-		{
-			step: deprovisioning.NewBTPOperatorCleanupStep(db.Operations(), provisionerClient, k8sClientProvider),
-		},
-		{
-			step: deprovisioning.NewAvsEvaluationsRemovalStep(avsDel, db.Operations(), externalEvalAssistant, internalEvalAssistant),
-		},
-		{
-			step:     deprovisioning.NewEDPDeregistrationStep(db.Operations(), edpClient, cfg.EDP),
-			disabled: cfg.EDP.Disabled,
-		},
-		{
-			step:     deprovisioning.NewIASDeregistrationStep(db.Operations(), bundleBuilder),
-			disabled: cfg.IAS.Disabled,
-		},
-		{
-			disabled: cfg.LifecycleManagerIntegrationDisabled,
-			step:     deprovisioning.NewDeleteKymaResourceStep(db.Operations(), cli, configProvider, cfg.KymaVersion),
-		},
-		{
-			disabled: cfg.LifecycleManagerIntegrationDisabled,
-			step:     deprovisioning.NewCheckKymaResourceDeletedStep(db.Operations(), cli),
-		},
-		{
-			disabled: cfg.ReconcilerIntegrationDisabled,
-			step:     deprovisioning.NewDeregisterClusterStep(db.Operations(), reconcilerClient),
-		},
-		{
-			disabled: cfg.ReconcilerIntegrationDisabled,
-			step:     deprovisioning.NewCheckClusterDeregistrationStep(db.Operations(), reconcilerClient, 90*time.Minute),
-		},
-		{
-			step: deprovisioning.NewRemoveRuntimeStep(db.Operations(), db.Instances(), provisionerClient, cfg.Provisioner.DeprovisioningTimeout),
-		},
-		{
-			step: deprovisioning.NewCheckRuntimeRemovalStep(db.Operations(), db.Instances(), provisionerClient),
-		},
-		{
-			step: deprovisioning.NewReleaseSubscriptionStep(db.Operations(), db.Instances(), accountProvider),
-		},
-		{
-			disabled: cfg.LifecycleManagerIntegrationDisabled,
-			step:     steps.DeleteKubeconfig(db.Operations(), cli),
-		},
-		{
-			step: deprovisioning.NewRemoveInstanceStep(db.Instances(), db.Operations()),
-		},
-	}
-	var stages []string
-	for _, step := range deprovisioningSteps {
-		if !step.disabled {
-			stages = append(stages, step.step.Name())
-		}
-	}
-	deprovisionManager.DefineStages(stages)
-	for _, step := range deprovisioningSteps {
-		if !step.disabled {
-			deprovisionManager.AddStep(step.step.Name(), step.step, nil)
-		}
-	}
-
-	queue := process.NewQueue(deprovisionManager, logs)
-	queue.Run(ctx.Done(), workersAmount)
-
-	return queue
-}
-
-func NewKymaOrchestrationProcessingQueue(ctx context.Context, db storage.BrokerStorage, runtimeOverrides upgrade_kyma.RuntimeOverridesAppender, provisionerClient provisioner.Client, pub event.Publisher, inputFactory input.CreatorForPlan, icfg *upgrade_kyma.TimeSchedule, pollingInterval time.Duration, runtimeVerConfigurator *runtimeversion.RuntimeVersionConfigurator, runtimeResolver orchestrationExt.RuntimeResolver, upgradeEvalManager *avs.EvaluationManager, cfg *Config, internalEvalAssistant *avs.InternalEvalAssistant, reconcilerClient reconciler.Client, notificationBuilder notification.BundleBuilder, logs logrus.FieldLogger, cli client.Client, speedFactor int) *process.Queue {
-
-	upgradeKymaManager := upgrade_kyma.NewManager(db.Operations(), pub, logs.WithField("upgradeKyma", "manager"))
-	upgradeKymaInit := upgrade_kyma.NewInitialisationStep(db.Operations(), db.Orchestrations(), db.Instances(),
-		provisionerClient, inputFactory, upgradeEvalManager, icfg, runtimeVerConfigurator, notificationBuilder)
-
-	upgradeKymaManager.InitStep(upgradeKymaInit)
-	upgradeKymaSteps := []struct {
-		disabled bool
-		weight   int
-		step     upgrade_kyma.Step
-		cnd      upgrade_kyma.StepCondition
-	}{
-		// check cluster configuration is the first step - to not execute other steps, when cluster configuration was applied
-		// this should be moved to the end when we introduce stages like in the provisioning process
-		// (also return operation, 0, nil at the end of apply_cluster_configuration)
-		{
-			weight:   1,
-			disabled: cfg.ReconcilerIntegrationDisabled,
-			step:     upgrade_kyma.NewCheckClusterConfigurationStep(db.Operations(), reconcilerClient, upgradeEvalManager, cfg.Reconciler.ProvisioningTimeout),
-			cnd:      upgrade_kyma.SkipForPreviewPlan,
-		},
-		{
-			weight: 1,
-			step:   steps.InitKymaTemplateUpgradeKyma(db.Operations()),
-		},
-		{
-			weight: 2,
-			step:   upgrade_kyma.NewGetKubeconfigStep(db.Operations(), provisionerClient),
-		},
-		{
-			weight:   2,
-			disabled: cfg.LifecycleManagerIntegrationDisabled,
-			step:     steps.SyncKubeconfigUpgradeKyma(db.Operations(), cli),
-		},
-		{
-			weight:   2,
-			disabled: cfg.LifecycleManagerIntegrationDisabled,
-			step:     upgrade_kyma.NewApplyKymaStep(db.Operations(), cli),
-		},
-		{
-			weight: 3,
-			cnd:    upgrade_kyma.WhenBTPOperatorCredentialsProvided,
-			step:   upgrade_kyma.NewBTPOperatorOverridesStep(db.Operations()),
-		},
-		{
-			weight: 4,
-			step:   upgrade_kyma.NewOverridesFromSecretsAndConfigStep(db.Operations(), runtimeOverrides, runtimeVerConfigurator),
-		},
-		{
-			weight: 8,
-			step:   upgrade_kyma.NewSendNotificationStep(db.Operations(), notificationBuilder),
-		},
-		{
-			weight:   10,
-			disabled: cfg.ReconcilerIntegrationDisabled,
-			step:     upgrade_kyma.NewApplyClusterConfigurationStep(db.Operations(), db.RuntimeStates(), reconcilerClient),
-			cnd:      upgrade_kyma.SkipForPreviewPlan,
-		},
-	}
-	for _, step := range upgradeKymaSteps {
-		if !step.disabled {
-			upgradeKymaManager.AddStep(step.weight, step.step, step.cnd)
-		}
-	}
-
-	orchestrateKymaManager := manager.NewUpgradeKymaManager(db.Orchestrations(), db.Operations(), db.Instances(),
-		upgradeKymaManager, runtimeResolver, pollingInterval, logs.WithField("upgradeKyma", "orchestration"),
-		cli, &cfg.OrchestrationConfig, notificationBuilder, speedFactor)
-	queue := process.NewQueue(orchestrateKymaManager, logs)
-
-	queue.Run(ctx.Done(), 3)
-
-	return queue
-}
-
-func NewClusterOrchestrationProcessingQueue(ctx context.Context, db storage.BrokerStorage, provisionerClient provisioner.Client,
-	pub event.Publisher, inputFactory input.CreatorForPlan, icfg *upgrade_cluster.TimeSchedule, pollingInterval time.Duration,
-	runtimeResolver orchestrationExt.RuntimeResolver, upgradeEvalManager *avs.EvaluationManager, notificationBuilder notification.BundleBuilder, logs logrus.FieldLogger,
-	cli client.Client, cfg Config, speedFactor int) *process.Queue {
-
-	upgradeClusterManager := upgrade_cluster.NewManager(db.Operations(), pub, logs.WithField("upgradeCluster", "manager"))
-	upgradeClusterInit := upgrade_cluster.NewInitialisationStep(db.Operations(), db.Orchestrations(), provisionerClient, inputFactory, upgradeEvalManager, icfg, notificationBuilder)
-	upgradeClusterManager.InitStep(upgradeClusterInit)
-
-	upgradeClusterSteps := []struct {
-		disabled  bool
-		weight    int
-		step      upgrade_cluster.Step
-		condition upgrade_cluster.StepCondition
-	}{
-		{
-			weight:    1,
-			step:      upgrade_cluster.NewLogSkippingUpgradeStep(db.Operations()),
-			condition: provisioning.DoForOwnClusterPlanOnly,
-		},
-		{
-			weight:    10,
-			step:      upgrade_cluster.NewSendNotificationStep(db.Operations(), notificationBuilder),
-			condition: provisioning.SkipForOwnClusterPlan,
-		},
-		{
-			weight:    10,
-			step:      upgrade_cluster.NewUpgradeClusterStep(db.Operations(), db.RuntimeStates(), provisionerClient, icfg),
-			condition: provisioning.SkipForOwnClusterPlan,
-		},
-	}
-
-	for _, step := range upgradeClusterSteps {
-		if !step.disabled {
-			upgradeClusterManager.AddStep(step.weight, step.step, step.condition)
-		}
-	}
-
-	orchestrateClusterManager := manager.NewUpgradeClusterManager(db.Orchestrations(), db.Operations(), db.Instances(),
-		upgradeClusterManager, runtimeResolver, pollingInterval, logs.WithField("upgradeCluster", "orchestration"),
-		cli, cfg.OrchestrationConfig, notificationBuilder, speedFactor)
-	queue := process.NewQueue(orchestrateClusterManager, logs)
-
-	queue.Run(ctx.Done(), 3)
-
-	return queue
 }
 
 func skipForPreviewPlan(operation internal.Operation) bool {
